COMPAT: boto import issues
Author: Jeff Reback <jeff@reback.net>

Closes #13136 from jreback/boto and squashes the following commits:

dcb02d2 [Jeff Reback] COMPAT: boto import issues
