+++ pandas/io/tests/test_data.py
@@ -472,9 +472,6 @@ class TestOptionsWarnings(tm.TestCase):
 
 
 class TestDataReader(tm.TestCase):
-    def test_is_s3_url(self):
-        from pandas.io.common import _is_s3_url
-        self.assertTrue(_is_s3_url("s3://pandas/somethingelse.com"))
 
     @network
     def test_read_yahoo(self):
